#2: refactor build and install scripts

+++ scripts/lib/Download-File.ps1
@@ -1,4 +1,4 @@
-param (
+﻿param (
     [Parameter(Mandatory = $true)]
     [string] $Url,
 
@@ -6,21 +6,38 @@ param (
     [string] $OutPath,
 
     [Parameter(Mandatory = $true)]
-    [string] $Sha256
+    [string] $Sha256,
+
+    [string] $Title
 )
 
 $ErrorActionPreference = 'Stop'
+
+if (-not $Title) {
+    $Title = $Url
+}
+
+Write-Host "Downloading $Title…"
 
 $outputDirectory = [IO.Path]::GetDirectoryName($OutPath)
 if (-not (Test-Path $outputDirectory)) {
     New-Item -ItemType Directory $outputDirectory
 }
 
-# TODO: Check hash and not redownload if already loaded
+function getHash {
+    (Get-FileHash $OutPath -Algorithm SHA256).Hash
+}
 
-Invoke-WebRequest $Url -OutFile $OutPath
-$hash = (Get-FileHash $OutPath -Algorithm SHA256).Hash
-if ($hash -ne $Sha256) {
-    Remove-Item $OutPath
-    throw "Invalid hash for $OutPath. Expected $Sha256, got $hash"
+if ((Test-Path $OutPath) -and (getHash -eq $Sha256)) {
+    Write-Host "$Title already loaded and have proper hash"
+    $false
+} else {
+    Invoke-WebRequest $Url -OutFile $OutPath
+    $hash = getHash
+    if ($hash -ne $Sha256) {
+        Remove-Item $OutPath
+        throw "Invalid hash for $Title. Expected $Sha256, got $hash"
+    }
+
+    $true
 }
